backend/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.19
asyncpg==0.28.1
alembic==1.11.1
pydantic==1.10.11
python-jose==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.24.1
cryptography==41.0.3
jinja2==3.1.2
python-dotenv==1.0.0
celery[redis]==5.3.1